动态展示菜单

+++ health_service/src/main/java/com/itheima/service/impl/MenuServiceImpl.java
@@ -1,12 +1,17 @@
 package com.itheima.service.impl;
 
 import com.alibaba.dubbo.config.annotation.Service;
+import com.itheima.constant.MessageConstant;
 import com.itheima.dao.MenuDao;
+import com.itheima.entity.Result;
 import com.itheima.pojo.Menu;
 import com.itheima.service.MenuService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.User;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -22,19 +27,70 @@ public class MenuServiceImpl implements MenuService {
     private MenuDao menuDao;
 
     /**
+     * 获取当前登录用户名
+     * @return
+     */
+    /*public String getUserName(){
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        return user.getUsername();
+    }*/
+
+    /**
      * 查询所有菜单项数据
      * @return
      */
     @Override
-    public List<Menu> findAll() {
-        //List<Menu> menuList = menuDao.findAll();
-        List<Menu> parentList = menuDao.findMenuByLevel1();
+    public List<Menu> findAll(String username) {
+        /*//List<Menu> menuList = menuDao.findAll();
+        List<Menu> parentList = menuDao.findMenuByLevel1(menuId);
         //遍历menuList
         for (Menu parent : parentList) {
             List<Menu> childrenList = menuDao.findMenuByLevel2AndParentId(parent.getId());
             parent.setChildren(childrenList);
         }
+        return parentList;*/
+        //获取菜单id集合
+        //获取菜单id集合
+        List<Integer> menuIdList = getMenuIdList(username);
+        List<Menu> parentList = new ArrayList<>();
+        for (Integer menuId : menuIdList) {
+            Menu level1Menu = menuDao.findMenuByLevel1(menuId);
+            if (level1Menu != null) {
+                parentList.add(level1Menu);
+            }
+        }
+        //遍历menuList
+        for (Menu parent : parentList) {
+            List<Menu> childrenList = new ArrayList<>();
+            for (Integer menuId : menuIdList) {
+                Menu level2Menu = menuDao.findMenuByLevel2AndParentId(parent.getId(),menuId);
+                if (level2Menu != null) {
+                    childrenList.add(level2Menu);
+                }
+            }
+            parent.setChildren(childrenList);
+        }
         return parentList;
+    }
+
+    /**
+     * 获取当前登录用户权限对应的菜单列表id集合
+     * @param username
+     * @return
+     */
+    private List<Integer> getMenuIdList(String username) {
+        Integer userId = menuDao.findUserIdByUsername(username);
+        Integer roleId = menuDao.findRoleIdByUserId(userId);
+        return menuDao.findMenuIdByRoleId(roleId);
+    }
+
+    /**
+     * 获取当前登录用户对应的角色id
+     * @return
+     */
+    private Integer getRoleId(String username) {
+        Integer userId = menuDao.findUserIdByUsername(username);
+        return menuDao.findRoleIdByUserId(userId);
     }
 
     /**
@@ -43,11 +99,22 @@ public class MenuServiceImpl implements MenuService {
      * @return
      */
     @Override
-    public void add(Menu menu) {
+    public Result add(Menu menu,String username) {
         String parentPath = getParentPath(menu);
-        //菜单中插入路径
-        menu.setPath(parentPath);
-        menuDao.add(menu);
+        //根据名称查询菜单表中是否已经有此菜单
+        Menu menuByName = menuDao.findByName(menu.getName());
+        if (menuByName != null) {
+            return new Result(false, MessageConstant.HAS_MENU);
+        }else {
+            //菜单中插入路径
+            menu.setPath(parentPath);
+            menuDao.add(menu);
+            //设置菜单-角色中间表关系
+            Integer id = menu.getId();
+            Menu addedMenu = menuDao.findByName(menu.getName());
+            menuDao.setRoleIdAndMenuId(getRoleId(username),addedMenu.getId());
+            return new Result(true,MessageConstant.ADD_MENU_SUCCESS);
+        }
     }
 
     /**
@@ -90,11 +157,18 @@ public class MenuServiceImpl implements MenuService {
      * @return
      */
     @Override
-    public void edit(Menu menu) {
+    public Result edit(Menu menu) {
         String parentPath = getParentPath(menu);
-        //菜单中插入路径
-        menu.setPath(parentPath);
-        menuDao.edit(menu);
+        //根据名称查询菜单表中是否已经有此菜单
+        Menu menuByName = menuDao.findByName(menu.getName());
+        if (menuByName != null) {
+            return new Result(false, MessageConstant.HAS_MENU);
+        }else {
+            //菜单中插入路径
+            menu.setPath(parentPath);
+            menuDao.edit(menu);
+            return new Result(true,MessageConstant.EDIT_MENU_SUCCESS);
+        }
     }
 
     /**
@@ -103,8 +177,16 @@ public class MenuServiceImpl implements MenuService {
      * @return
      */
     @Override
-    public void deleteById(Integer id) {
+    public void deleteById(Integer id){
+        //查询当前删除菜单是否有子菜单
+        Integer count = menuDao.findCountOfChildrentMenu(id);
+        if (count > 0){
+            throw new RuntimeException(MessageConstant.DELETE_MENU_FAIL2);
+        }
+        //删除菜单-角色中间表关系
+        menuDao.deleteRoleIdAndMenuId(id);
         menuDao.deleteById(id);
+        //todo
     }
 
     /**
@@ -112,16 +194,87 @@ public class MenuServiceImpl implements MenuService {
      * @return
      */
     @Override
-    public List<Menu> getMenuList() {
-        List<Menu> parentList = menuDao.findMenuByLevel1();
+    public List<Menu> getMenuList(String username) {
+        //获取菜单id集合
+        List<Integer> menuIdList = getMenuIdList(username);
+        List<Menu> parentList = new ArrayList<>();
+        for (Integer menuId : menuIdList) {
+            Menu level1Menu = menuDao.findMenuByLevel1(menuId);
+            if (level1Menu != null) {
+                parentList.add(level1Menu);
+            }
+        }
         //遍历menuList
         for (Menu parent : parentList) {
             parent.setTitle(parent.getName());
-            List<Menu> childrenList = menuDao.findMenuByLevel2AndParentId(parent.getId());
+            List<Menu> childrenList = new ArrayList<>();
+            for (Integer menuId : menuIdList) {
+                Menu level2Menu = menuDao.findMenuByLevel2AndParentId(parent.getId(),menuId);
+                if (level2Menu != null){
+                    childrenList.add(level2Menu);
+                }
+            }
             for (Menu children : childrenList) {
                 children.setTitle(children.getName());
             }
             parent.setChildren(childrenList);
+        }
+        return parentList;
+    }
+
+
+
+
+    /**
+     * 菜单项分页查询
+     * @param queryString
+     * @return
+     */
+    @Override
+    public List<Menu> findPage(String queryString,String username) {
+        /*List<Menu> parentList = menuDao.findMenuByLevel1AndCondition(queryString);
+        //遍历parentList
+        if (parentList.size()>0) {
+            for (Menu parent : parentList) {
+                //List<Menu> childrenList = menuDao.findMenuByLevel2AndParentIdAndCondition(parent.getId(),queryString);
+                List<Menu> childrenList = menuDao.findMenuByLevel2AndParentId(parent.getId());
+                parent.setChildren(childrenList);
+            }
+        }else {
+            parentList = menuDao.findMenuByLevel2AndCondition(queryString);
+        }
+        return parentList;*/
+        //获取菜单id集合
+        List<Integer> menuIdList = getMenuIdList(username);
+        //获取一级菜单集合
+        List<Menu> parentList = new ArrayList<>();
+        for (Integer menuId : menuIdList) {
+            Menu level1Menu = menuDao.findMenuByLevel1AndCondition(queryString,menuId);
+            if (level1Menu != null) {
+                parentList.add(level1Menu);
+            }
+        }
+        //遍历parentList
+        if (parentList.size()>0) {
+            for (Menu parent : parentList) {
+                parent.setTitle(parent.getName());
+                List<Menu> childrenList = new ArrayList<>();
+                for (Integer menuId : menuIdList) {
+                    Menu level2Menu = menuDao.findMenuByLevel2AndParentId(parent.getId(), menuId);
+                    if (level2Menu != null){
+                        childrenList.add(level2Menu);
+                    }
+                }
+                for (Menu children : childrenList) {
+                    children.setTitle(children.getName());
+                }
+                parent.setChildren(childrenList);
+            }
+        }else {
+            for (Integer menuId : menuIdList) {
+                Menu level2Menu = menuDao.findMenuByLevel2AndCondition(queryString,menuId);
+                parentList.add(level2Menu);
+            }
         }
         return parentList;
     }
